Extract variable display functions to utility module
